Move BibDatabaseContextChangedEvent to net.sf.jabref.model.database.event
