--- src/main/java/net/sf/jabref/event/BibDatabaseContextChangedEvent.java
@@ -1,10 +1,0 @@
-package net.sf.jabref.event;
-
-import net.sf.jabref.model.metadata.event.MetaDataChangedEvent;
-
-/**
- * This Event is automatically fired in the same time as {@link EntryEvent}, {@link GroupUpdatedEvent} or {@link MetaDataChangedEvent}.
- */
-public class BibDatabaseContextChangedEvent {
-    // no data
-}
