+++ src/main/java/net/sf/jabref/logic/autosave/BackupManager.java
@@ -17,12 +17,12 @@ import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
-import net.sf.jabref.event.BibDatabaseContextChangedEvent;
 import net.sf.jabref.logic.exporter.BibtexDatabaseWriter;
 import net.sf.jabref.logic.exporter.FileSaveSession;
 import net.sf.jabref.logic.exporter.SaveException;
 import net.sf.jabref.logic.exporter.SavePreferences;
 import net.sf.jabref.model.database.BibDatabaseContext;
+import net.sf.jabref.model.database.event.BibDatabaseContextChangedEvent;
 import net.sf.jabref.preferences.JabRefPreferences;
 
 import com.google.common.eventbus.Subscribe;
