+++ src/main/java/net/sf/jabref/model/database/event/BibDatabaseContextChangedEvent.java
@@ -1,0 +1,10 @@
+package net.sf.jabref.model.database.event;
+
+import net.sf.jabref.model.metadata.event.MetaDataChangedEvent;
+
+/**
+ * This Event is automatically fired at the same time as {@link EntryEvent}, {@link GroupUpdatedEvent} or {@link MetaDataChangedEvent}.
+ */
+public class BibDatabaseContextChangedEvent {
+    // no data
+}
